+++ changelog.md
@@ -95,6 +95,7 @@
 - Added new boxes components for info and links
 - Revamped kamehouse homepage
 - Added custom icons support on crud manager ui
+- Added info image component on crud manager ui
 - Fixed bugs
 
 #### v8.12